+++ RELEASES.md
@@ -5,6 +5,19 @@ modification, are permitted in any medium without royalty provided the
 copyright notice and this notice are preserved. This file is offered
 as-is, without any warranty. -->
 
+Version 1.7.0 (unreleased)
+==========================
+
+  * For the experimental feature [`num-traits`][feat-exp-1-7], the
+    following traits were implemented where applicable:
+      * [`OverflowingAdd`][nt-0-2-oa], [`OverflowingSub`][nt-0-2-os],
+        [`OverflowingMul`][nt-0-2-om]
+
+[feat-exp-1-7]: https://tspiteri.gitlab.io/fixed/dev/fixed/#experimental-optional-features
+[nt-0-2-oa]: https://docs.rs/num-traits/^0.2/num_traits/ops/overflowing/trait.OverflowingAdd.html
+[nt-0-2-om]: https://docs.rs/num-traits/^0.2/num_traits/ops/overflowing/trait.OverflowingMul.html
+[nt-0-2-os]: https://docs.rs/num-traits/^0.2/num_traits/ops/overflowing/trait.OverflowingSub.html
+
 Version 1.6.0 (2021-02-05)
 ==========================